# subscriptions.swagger.yaml
openapi: 3.0.3
info:
  title: Subscriptions API
  version: "1.0.0"
  description: |
    Subscriptions-focused API specification for the BasicDiet backend.

    This document covers:
    - App / mobile subscription endpoints under `/api/subscriptions/*`
    - Admin / dashboard subscription endpoints under `/api/admin/*`
    - Dashboard aliases under `/api/dashboard/*`

    Authentication:
    - `GET /subscriptions/menu` is public.
    - All other `/subscriptions/*` endpoints require an App bearer token.
    - All `/admin/*` and `/dashboard/*` endpoints require a Dashboard bearer token.

    Localization:
    - Pass `?lang=en` or `?lang=ar`
    - Or send `Accept-Language: en|ar`
    - Precedence: query string > header > fallback
    - Fallback language: `ar`

    Response envelope:
    - Most endpoints return `{ "status": true, "data": ... }`
    - Errors return `{ "status": false, "error": { "code": "...", "message": "..." } }`
    - A few legacy or development-only routes may return a `message` field instead of a full `data` object

    Important frontend rules:
    - Never use localized `*Label` fields for business logic
    - Always use machine fields such as `status`, `paymentStatus`, `walletType`, `source`, and `error.code`
    - Always call the relevant `verify-payment` endpoint after payment-provider redirects when such an endpoint exists
    - Always send an `idempotencyKey` for checkout creation and reuse it on safe checkout retries
...
servers:
...
    description: Local
...
components:
  securitySchemes:
    AppBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: App user bearer token
    DashboardBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard bearer token
  parameters:
    LangQuery:
      name: lang
      in: query
      required: false
      schema:
        type: string
        enum: [ar, en]
      description: Optional language override. Takes precedence over `Accept-Language`.
...
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        status:
          type: boolean
          example: true
        data:
          type: object
      required: [status]
...
paths:
  /subscriptions/menu:
    get:
      tags: [Menu]
      summary: Get subscription menu
      description: Public catalog endpoint used to build the subscription purchase UI.
      parameters:
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Subscription catalog returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /subscriptions:
    get:
      tags: [Subscriptions]
      summary: List current user subscriptions
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Subscriptions returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /subscriptions/quote:
    post:
      tags: [Checkout]
      summary: Quote subscription checkout
      description: Calculates the subscription price before checkout creation.
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LangQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequestBody'
      responses:
        '200':
          description: Quote calculated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: VALIDATION_ERROR or INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/checkout:
    post:
      tags: [Checkout]
      summary: Create checkout draft
      description: |
        Creates a checkout draft and payment intent.
        Send `idempotencyKey` through `Idempotency-Key`, `X-Idempotency-Key`, or `body.idempotencyKey`.
      security:
        - AppBearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/QuoteRequestBody'
                - type: object
                  properties:
                    idempotencyKey:
                      type: string
                    successUrl:
                      type: string
                    backUrl:
                      type: string
      responses:
        '201':
          description: Checkout draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '200':
          description: Existing reusable draft returned for the same idempotency key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: VALIDATION_ERROR or INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: IDEMPOTENCY_CONFLICT or CHECKOUT_IN_PROGRESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/checkout-drafts/{draftId}:
    get:
      tags: [Checkout]
      summary: Get checkout draft status
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Draft status returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/checkout-drafts/{draftId}/verify-payment:
    post:
      tags: [Checkout]
      summary: Verify checkout draft payment
      description: Always call this after payment-provider redirects for subscription checkout.
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DraftId'
      responses:
        '200':
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID or INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: CHECKOUT_IN_PROGRESS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: CONFIG
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: PAYMENT_PROVIDER_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/activate:
    post:
      tags: [Checkout]
      summary: Development-only activate route
      description: Present only when `NODE_ENV !== production`.
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '200':
          description: Development response
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                  message:
                    type: string
  /subscriptions/{id}/renewal-seed:
    get:
      tags: [Subscriptions]
      summary: Get renewal seed
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Renewal seed returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RENEWAL_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/renew:
    post:
      tags: [Subscriptions]
      summary: Renew subscription
      description: |
        Starts a renewal checkout flow from the current subscription.
        Use `idempotencyKey` for safe retries.
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenewRequestBody'
      responses:
        '201':
          description: Renewal checkout created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '200':
          description: Existing reusable checkout returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID or INVALID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: IDEMPOTENCY_CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: RENEWAL_PREMATURE or RENEWAL_UNAVAILABLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}:
    get:
      tags: [Subscriptions]
      summary: Get subscription by id
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Subscription returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/timeline:
    get:
      tags: [Timeline & Days]
      summary: Get subscription timeline
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Timeline returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/days:
    get:
      tags: [Timeline & Days]
      summary: Get all subscription days
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Days returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/today:
    get:
      tags: [Timeline & Days]
      summary: Get today's subscription day
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Today returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/days/{date}:
    get:
      tags: [Timeline & Days]
      summary: Get subscription day by date
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/DateParam'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Day returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/days/{date}/selection:
    put:
      tags: [Day Planning]
      summary: Save day selection
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/DateParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DaySelectionRequest'
      responses:
        '200':
          description: Selection saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID, INVALID_DATE, DAILY_CAP, or validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: LOCKED or DATA_INTEGRITY_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: SUB_INACTIVE or SUB_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/days/{date}/confirm:
    post:
      tags: [Day Planning]
      summary: Confirm day planning
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/DateParam'
      responses:
        '200':
          description: Day confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_DATE or LOCKED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: CANONICAL_DAY_PLANNING_DISABLED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: PLANNING_INCOMPLETE, PREMIUM_OVERAGE_PAYMENT_REQUIRED, ONE_TIME_ADDON_PAYMENT_REQUIRED, SUB_INACTIVE, or SUB_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/days/{date}/pickup/prepare:
    post:
      tags: [Day Planning]
      summary: Prepare pickup day
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/DateParam'
      responses:
        '200':
          description: Pickup prepared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID, INVALID_DATE, LOCKED, or INSUFFICIENT_CREDITS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: INVALID_TRANSITION
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: SUB_INACTIVE or SUB_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/days/{date}/custom-salad:
    post:
      tags: [Day Planning]
      summary: Create custom salad payment
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/DateParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ingredients]
              properties:
                ingredients:
                  type: array
                  items:
                    type: object
                    properties:
                      ingredientId:
                        type: string
                      quantity:
                        type: integer
                successUrl:
                  type: string
                backUrl:
                  type: string
      responses:
        '200':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID, INVALID_DATE, LOCKED, or MAX_EXCEEDED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: SUB_INACTIVE or SUB_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/days/{date}/custom-meal:
    post:
      tags: [Day Planning]
      summary: Create custom meal payment
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/DateParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ingredients]
              properties:
                ingredients:
                  type: array
                  items:
                    type: object
                    properties:
                      ingredientId:
                        type: string
                      quantity:
                        type: integer
                successUrl:
                  type: string
                backUrl:
                  type: string
      responses:
        '200':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID, INVALID_DATE, LOCKED, or MAX_EXCEEDED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: SUB_INACTIVE or SUB_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /subscriptions/{id}/wallet:
    get:
      tags: [Premium & Payments]
      summary: Get subscription wallet
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Wallet snapshot returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /subscriptions/{id}/wallet/history:
    get:
      tags: [Premium & Payments]
      summary: Get wallet history
      security:
        - AppBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
        - $ref: '#/components/parameters/LangQuery'
      responses:
        '200':
          description: Wallet history returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: INVALID_ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...